backend/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
EasyProcess==1.1
entrypoint2==1.1
MouseInfo==0.1.3
mss==9.0.1
Pillow==10.1.0
psycopg2-binary==2.9.9
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python3-xlib==0.15
pytweening==1.0.7
sqlparse==0.4.4
typing_extensions==4.8.0
